+++ starters/nextjs-starter-approuter-ts/components/article-view.tsx
@@ -1,13 +1,17 @@
 "use client";
 
-import { useArticle } from "@pantheon-systems/pcc-react-sdk";
+import { findTab, useArticle } from "@pantheon-systems/pcc-react-sdk";
 import type { Article, PublishingLevel } from "@pantheon-systems/pcc-react-sdk";
 import { ArticleRenderer } from "@pantheon-systems/pcc-react-sdk/components";
 import Image from "next/image";
 import Link from "next/link";
-import React from "react";
-import { getSeoMetadata } from "../lib/utils";
+import React, { useMemo } from "react";
+import toast, { Toaster } from "react-hot-toast";
+import HeaderLink from "../assets/icons/HeaderLink";
+import { getSeoMetadata, parseAsTabTree } from "../lib/utils";
 import { clientSmartComponentMap } from "./smart-components/client-components";
+import { TableOfContents } from "./table-of-contents";
+import { useSearchParams } from "next/navigation";
 
 const ELEMENT_STYLES_TO_OVERRIDE = [
   /fontSize/,
@@ -52,6 +56,7 @@ const componentOverrideMap = {
 type ArticleViewProps = {
   article: Article;
   onlyContent?: boolean;
+  tabId?: string | null;
   publishingLevel: keyof typeof PublishingLevel;
 };
 
@@ -69,7 +74,7 @@ const ArticleHeader = ({
   if (!author?.name && !article.updatedAt) return null;
 
   return (
-    <div className="border-b-base-300 text-neutral-content mb-14 mt-6 flex w-full flex-row gap-x-4 border-b-[1px] py-4">
+    <div className="border-y-base-300 text-neutral-content mb-14 mt-6 flex w-full flex-row gap-x-4 border-y-[1px] py-4">
       {author?.name ? (
         <>
           <div>
@@ -109,22 +114,63 @@ const ArticleHeader = ({
 export function StaticArticleView({
   article,
   onlyContent,
-}: Pick<ArticleViewProps, "article" | "onlyContent">) {
+  tabId
+}: Pick<ArticleViewProps, "article" | "onlyContent" | "tabId">) {
   const seoMetadata = getSeoMetadata(article);
+
+  const tabTree =
+    article.resolvedContent == null
+      ? null
+      : parseAsTabTree(article.resolvedContent);
+
+  const currentTab =
+    tabTree != null && tabId != null ? findTab(tabTree, tabId) : null;
 
   return (
     <div className="px-8 lg:px-4">
       <ArticleHeader article={article} seoMetadata={seoMetadata} />
-      <ArticleRenderer
-        article={article}
-        componentMap={componentOverrideMap}
-        smartComponentMap={clientSmartComponentMap}
-        __experimentalFlags={{
-          disableAllStyles: !!onlyContent,
-          preserveImageStyles: true,
-          useUnintrusiveTitleRendering: true,
-        }}
-      />
+
+      <div className="flex justify-start gap-x-[50px] lg:gap-x-[115px]">
+        {article.renderAsTabs && tabTree ? (
+          <TableOfContents tabTree={tabTree} activeTab={tabId} />
+        ) : null}
+
+        <div className="flex-1">
+          {currentTab?.tabProperties?.title ? (
+            <h3 className="my-0 flex items-center gap-x-4">
+              <span>{currentTab.tabProperties.title}</span>
+              <Link
+                href={
+                  typeof window === "undefined" ? "#" : window.location.href
+                }
+                aria-label={`Link to "${currentTab.tabProperties.title}"`}
+                onClick={() => {
+                  navigator.clipboard.writeText(window.location.href);
+                  toast.success(
+                    "A link to this page has been copied to your clipboard",
+                    {
+                      position: "top-right",
+                    },
+                  );
+                }}
+              >
+                <HeaderLink height={25} width={25} />
+              </Link>
+            </h3>
+          ) : null}
+          <ArticleRenderer
+            article={article}
+            tabId={tabId}
+            componentMap={componentOverrideMap}
+            smartComponentMap={clientSmartComponentMap}
+            __experimentalFlags={{
+              disableAllStyles: !!onlyContent,
+              preserveImageStyles: true,
+              useUnintrusiveTitleRendering: true,
+            }}
+          />
+        </div>
+      </div>
 
       <div className="border-base-300 mt-16 flex w-full flex-wrap gap-x-3 gap-y-3 border-t-[1px] pt-9 lg:mt-32">
         {seoMetadata.keywords != null
@@ -148,8 +194,12 @@ export function StaticArticleView({
 export default function ArticleView({
   article,
   onlyContent,
+  tabId,
   publishingLevel,
 }: ArticleViewProps) {
+  const searchParams = useSearchParams();
+  const currentTabId = useMemo(() => searchParams.get("tabId") || tabId, [searchParams, tabId]);
+
   const { data } = useArticle(
     article.id,
     {
@@ -164,6 +214,15 @@ export default function ArticleView({
   const hydratedArticle = data?.article ?? article;
 
   return (
-    <StaticArticleView article={hydratedArticle} onlyContent={onlyContent} />
+    <>
+      <div>
+        <Toaster />
+      </div>
+      <StaticArticleView
+        article={hydratedArticle}
+        onlyContent={onlyContent}
+        tabId={currentTabId}
+      />
+    </>
   );
 }